requirements.txt
requests
azure.functions